requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
crcmod==1.7
cryptography==45.0.6
cycler==0.12.1
deformable-attention==0.0.20
distro==1.9.0
einops==0.8.1
filelock==3.14.0
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2024.6.1
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.8
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
Jinja2==3.1.4
jiter==0.11.1
jmespath==0.10.0
joblib==1.5.1
kiwisolver==1.4.9
lazy_loader==0.4
lightning-utilities==0.15.2
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.5
mdurl==0.1.2
model-index==0.1.11
modelscope==1.29.1
mpmath==1.3.0
multidict==6.6.4
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==2.6.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.3.2
pillow==11.0.0
polars==1.32.3
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.3
pytz==2023.4
PyYAML==6.0.2
regex==2025.7.34
requests==2.28.2
rich==13.4.2
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
segmentation_models_pytorch==0.5.0
setuptools==80.9.0
simsimd==6.5.1
six==1.17.0
sniffio==1.3.1
stringzilla==3.12.6
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.6.11
timm==1.0.19
tokenizers==0.21.4
torch==2.8.0+cu126
torchmetrics==1.8.1
torchvision==0.23.0+cu126
tqdm==4.65.2
transformers @ file:///data/stone/segmentation/transformers
triton==3.4.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.189
ultralytics-thop==2.0.16
urllib3==1.26.20
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.20.1
